[GYJB-162] Content 추상 클래스를 만들고 Link 자식 클래스 생성
