+++ src/main/java/com/linklip/linklipserver/domain/Content.java
@@ -2,7 +2,6 @@ package com.linklip.linklipserver.domain;
 
 import javax.persistence.*;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.Where;
@@ -12,22 +11,17 @@ import org.hibernate.annotations.Where;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 // 자동으로 where절에 deleted = false 이라는 SQL구문을 추가
 @Where(clause = "is_deleted = false")
-public class Content extends JpaBaseDomain {
+@DiscriminatorColumn(name = "type")
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+public abstract class Content extends JpaBaseDomain {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "content_id")
     private Long id;
 
-    @Column(nullable = false)
-    private String linkUrl;
-
-    private String linkImg;
-
-    private String title;
-
-    @Lob // column type을 longtext로 설정
-    private String text;
+    @Column(insertable = false, updatable = false) // 읽기 전용으로 선언
+    private String type;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "category_id")
@@ -36,17 +30,7 @@ public class Content extends JpaBaseDomain {
     @Column(nullable = false)
     private boolean isDeleted;
 
-    @Builder
-    public Content(String linkUrl, String linkImg, String title, String text, Category category) {
-        this.linkUrl = linkUrl;
-        this.linkImg = linkImg;
-        this.title = title;
-        this.text = text;
-        this.category = category;
-    }
-
-    public void update(String title, Category category) {
-        this.title = title;
+    public void updateCategory(Category category) {
         this.category = category;
     }
